+++ cli-tools/cli/commands/upgrade-project/upgrade-config.js
@@ -3,7 +3,7 @@ const path = require('path');
 module.exports = [
   {
     from: '1.x',
-    to: 'latest v2',
-    path: path.join(__dirname, 'codemod', '2.x')
+    to: 'v2.0.0',
+    path: path.join(__dirname, 'codemod', '2.0.0')
   }
 ];